+++ eq-author/src/components/Answers/BasicAnswer/index.js
