+++ eq-author/src/components/Answers/CurrencyAnswer/CurrencyAnswer.test.js
