+++ eq-author/src/components/Answers/CurrencyAnswer/index.js
