+++ eq-author/src/components/Answers/Dummy/MultipleChoice/index.js
