+++ eq-author/src/components/Answers/MultipleChoiceAnswer/MultipleChoiceAnswer.test.js
