+++ eq-author/src/components/Answers/MultipleChoiceAnswer/Option.js
