+++ eq-author/src/components/Answers/MultipleChoiceAnswer/OptionTransition.js
