+++ eq-author/src/components/Answers/MultipleChoiceAnswer/index.js
