+++ eq-author/src/containers/App/index.js
@@ -12,9 +12,10 @@ import PrivateRoute from "components/PrivateRoute";
 import RedirectRoute from "components/RedirectRoute";
 import { Routes as RoutePaths } from "utils/UrlUtils";
 
-import QuestionnairesPage from "components/QuestionnairesPage";
 import SignInPage from "containers/SignInPage";
 import QuestionnaireDesignPage from "features/design";
+import QuestionnairesPage from "features/dashboard";
+
 import NotFoundPage from "containers/NotFoundPage";
 import { Provider, connect } from "react-redux";
 import { isSignedIn } from "redux/auth/reducer";
